+++ test/jsRemote/jsVrpcRemoteTest.js
@@ -20,8 +20,8 @@ describe('An instance of the VrpcRemote class', () => {
     vrpc = new VrpcRemote({ topicPrefix: 'vrpc_test' })
     assert.ok(vrpc)
   })
-  it('should return available classes and functions', async () => {
-    await new Promise(resolve => setTimeout(resolve, 500))
+  it.skip('should return available classes and functions', async () => {
+    await new Promise(resolve => setTimeout(resolve, 300))
     console.log('Agents:', await vrpc.getAvailableAgents())
     console.log('Classes:', await vrpc.getAvailableClasses())
     console.log('Member Functions:', await vrpc.getAvailableMemberFunctions('js', 'TestClass'))